health(criteria): #80 split up parse-token specs & some spec file renaming
