+++ src/criteria/test/parsing/parse-criteria.spec.ts
@@ -1,7 +1,7 @@
 import { parseCriteria } from "../../parser";
 import { and, inRange, inSet, notInSet, or, ValueCriterion } from "../../value-criterion";
 
-describe("criteria-parser", () => {
+describe("parse-criteria", () => {
     function shouldParse(stringified: string, expected: ValueCriterion): void {
         it(`should parse ${stringified}`, () => {
             const parse = () => parseCriteria(stringified);
